backend/requirements.txt
fastapi
uvicorn
langchain-core>=0.2.20
langchain-community>=0.2.10
langchain-openai>=0.1.15
langsmith
pydantic
pinecone
python-dotenv
requests
openai